requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-DebugToolbar==0.16.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.0.0
PyJWT==2.10.1
pymongo==4.10.1
pyotp==2.9.0
python-dotenv==1.2.1
qrcode==8.2
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.4
WTForms==3.2.1